finished ecdsa refactor
